fix(depsync): ensure fresh main branch when Dagger reuses cached directory
- Add git checkout main and git pull to CloneRepo function
- Handle case where Dagger reuses old cached directory
- Ensure repository is always on latest main branch
- Fix unused parameter warning in handlePRConflicts function
