+++ pkg/config/config.go
@@ -14,8 +14,9 @@ type GitConfig struct {
 }
 
 type Config struct {
-	Repositories []string  `mapstructure:"repositories"`
-	Git          GitConfig `mapstructure:"git"`
+	Repositories        []string  `mapstructure:"repositories"`
+	Git                 GitConfig `mapstructure:"git"`
+	DeleteConflictedPRs bool      `mapstructure:"delete_conflicted_prs"`
 }
 
 func Load(configPath string) (*Config, error) {
@@ -30,6 +31,11 @@ func Load(configPath string) (*Config, error) {
 
 	if err := viper.Unmarshal(&config); err != nil {
 		return nil, err
+	}
+
+	// Set default value for DeleteConflictedPRs if not specified
+	if !viper.IsSet("delete_conflicted_prs") {
+		config.DeleteConflictedPRs = true
 	}
 
 	return &config, nil
